+++ src/main/recorder.ts
@@ -35,7 +35,7 @@ type RecorderOptionsType = {
      */
     constructor(options: RecorderOptionsType) {
         this._options = options;
-        console.debug("[Recorder] Construcing recorder with: ", this._options);
+        console.debug("[Recorder] Constructing recorder with: ", this._options);
 
         if (!fs.existsSync(this._options.bufferStorageDir)) {
             console.log("[Recorder] Creating dir:", this._options.bufferStorageDir);