+++ src/settings/AdvancedSettings.tsx
@@ -3,6 +3,11 @@ import { openDirectorySelectorDialog } from './settingUtils';
 import TextField from '@mui/material/TextField';
 import Stack from '@mui/material/Stack';
 import ConfigContext from "./ConfigContext";
+import Box from '@mui/material/Box';
+import InfoIcon from '@mui/icons-material/Info';
+import IconButton from '@mui/material/IconButton';
+import Tooltip from '@mui/material/Tooltip';
+import { configSchema } from '../main/configSchema'
 
 const ipc = window.electron.ipcRenderer;
 
@@ -25,6 +30,7 @@ export default function GeneralSettings() {
 
 
   const style = {
+    width: '405px',
     "& .MuiOutlinedInput-root": {
       "&.Mui-focused fieldset": {borderColor: "#bb4220"},
       "& > fieldset": {borderColor: "black" }
@@ -37,34 +43,48 @@ export default function GeneralSettings() {
     <Stack
       component="form"
       sx={{
-        '& > :not(style)': { m: 1, width: '50ch' },
+        '& > :not(style)': { m: 0, width: '50ch' },
       }}
       noValidate
       autoComplete="off"
     >
-      <TextField 
-        value={config.bufferStoragePath}
-        id="buffer-path" 
-        label="Buffer Path" 
-        variant="outlined" 
-        onClick={() => openDirectorySelectorDialog('bufferStoragePath')} 
-        InputLabelProps={{ shrink: true }}
-        sx={style}
-        inputProps={{ style: { color: "white" } }}
-      />
-      <TextField 
-        value={config.minEncounterDuration}
-        onChange={event => { modifyConfig("minEncounterDuration", parseInt(event.target.value, 10)) }}
-        id="max-storage" 
-        label="Min Encounter Duration (sec)" 
-        variant="outlined" 
-        type="number" 
-        error= { config.minEncounterDuration < 1 }
-        helperText={(config.minEncounterDuration < 1) ? "Must be positive" : ' '}
-        InputLabelProps={{ shrink: true }}
-        sx={style}
-        inputProps={{ style: { color: "white" } }}
-      />
+      <Box component="span" sx={{ display: 'flex', alignItems: 'center'}}>
+        <TextField 
+          value={config.bufferStoragePath}
+          id="buffer-path" 
+          label="Buffer Path" 
+          variant="outlined" 
+          onClick={() => openDirectorySelectorDialog('bufferStoragePath')} 
+          InputLabelProps={{ shrink: true }}
+          sx={{...style, my: 1}}
+          inputProps={{ style: { color: "white" } }}
+        />
+        <Tooltip title={configSchema["bufferStoragePath"].description} >
+          <IconButton>
+            <InfoIcon style={{ color: 'white' }}/>
+          </IconButton>
+        </Tooltip>
+      </Box>
+      <Box component="span" sx={{ display: 'flex', alignItems: 'center'}}>
+        <TextField 
+          value={config.minEncounterDuration}
+          onChange={event => { modifyConfig("minEncounterDuration", parseInt(event.target.value, 10)) }}
+          id="max-storage" 
+          label="Min Encounter Duration (sec)" 
+          variant="outlined" 
+          type="number" 
+          error= { config.minEncounterDuration < 1 }
+          helperText={(config.minEncounterDuration < 1) ? "Must be positive" : ' '}
+          InputLabelProps={{ shrink: true }}
+          sx={{...style, my: 1}}
+          inputProps={{ style: { color: "white" } }}
+        />
+        <Tooltip title={configSchema["minEncounterDuration"].description} sx={{position: 'fixed', left: '572px', top: '140px'}} >
+          <IconButton>
+            <InfoIcon style={{ color: 'white' }}/>
+          </IconButton>
+        </Tooltip>
+      </Box>
     </Stack>
   );
 }