+++ src/settings/Settings.tsx
@@ -32,7 +32,7 @@ function TabPanel(props: TabPanelProps) {
       {...other}
     >
       {value === index && (
-        <Box sx={{ p: 3 }}>
+        <Box sx={{ p: 2 }}>
           <Typography component={'span'}>{children}</Typography>
         </Box>
       )}